requirements.txt
attrs==24.2.0
boto3==1.35.20
botocore==1.35.20
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
h11==0.14.0
idna==3.10
japanize-matplotlib==1.1.3
jmespath==1.0.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
s3transfer==0.10.2
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
